docs: log signup form unification
