+++ app/inscricoes/lider/[liderId]/evento/[eventoId]/page.tsx
