+++ components/organisms/index.ts
@@ -8,3 +8,4 @@ export { default as BlogPostsList } from './BlogPostsList'
 export { default as FormWizard } from './FormWizard'
 export { default as InscricaoWizard } from './InscricaoWizard'
 export { default as InscricaoLojaWizard } from './InscricaoLojaWizard'
+export { default as EventForm } from './EventForm'
